Add EGS_AEnvelope & EGS_ASwitchedEnvelope geoms
This is an efficient envelope geometry based on the EGS_FastEnvelope
geometry.  The defining feature of EGS_AEnvelope is that base geometry
regions containing inscribed geometries are discovered automatically
using a Monte Carlo simulation.

EGS_ASwitched Envelope allows user codes to enable/disable one or more
inscribed geometries (useful for e.g. stepped HDR sources).

Distributed with this library are the files sobol.h, sobol.cpp,
gzstream.h & gzstream.cpp all of which are GPL licensed (see
files for license details).

The copyright for this code is transferred to the
National Research Council of Canada