Changing egs++ geometry getMass to getVolume
The egs_base_geometry getMass virtual function is exchanged with
a getVolume function that returns region volume instead of mass.
This change was performed because typically, initGeometry() is
invoked before any media data is read in egs_applications, thus
getMass' dependence on medium data could be unreliable.  The
getMass function was thus replaced with a getVolume function,
which only relies on internal gemoetry definitions to return a
value.
The geometries in egs_autoenvelope, egs_nd_geometry, egs_rz,
and egs_spheres are affected.  The ausgab dose scoring object
which outputs a 3ddose file was also changed to reflect the
change from getMass to getVolume.  This modification of
egs_dose_scoring also now includes relativeRho in the mass
calculation.