Adding egs_mird and relevant source classes
The application egs_mird was built as a collaboration between
Ting-Yim Lee at the Robarts Research Institute and Rowan Thomson
at the Carleton Laboratory for Radiotherapy Physics to perform
fast 3ddose calculations of targeted radionuclide therapy using a
very simple interface for increased ease of use.  In addition to
the egs_mird application, a new source called egs_internal_source
is added that allows a user to generate sources throughout an XYZ
phantom using a list of voxel regions and weights.  Additionally,
egs_radionuclide_source has been modified to allow any other
source as an input (and can thus be combined with the new internal
source).